Refactor SantoDelGiornoController and routes
Modified several methods in SantoDelGiornoController to improve handling of Santo data. In particular, the 'show' method was adjusted to accept an integer ID instead of a Santo object and added error checking to handle non-existent Santo. The 'findByDate' and 'findByName' methods have been updated to provide a more descriptive message. Correspondingly, adjusted the route URL in api_v1.php to reflect the 'show' method changes.