Update template structure and introduce new features
This commit primarily updates the template structure to incorporate a new navbar and footer. It also includes the creation of a documentation markdown file and the addition of new views and routes. Extra dependencies related to typography in Tailwind CSS are also included in this commit to provide better typographic control.