Add 'findOnomastico' function and route
A new function named 'findOnomastico' was added to SantoDelGiornoController for finding saints names related to a given name. Furthermore, a route for accessing this function was included in 'api_v1.php'. The logic includes searching for names matching or containing the input name with either 'onomastico' or 'onomastico_secondario' status set to true.